requirements.txt
asgiref==3.6.0
cffi==1.15.1
cryptography==41.0.1
Django==4.1.7
django-cors-headers==3.14.0
django-encrypted-id==0.3.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
hashids==1.3.1
mysqlclient==2.1.1
Pillow==9.5.0
pycparser==2.21
pycrypto==2.6.1
PyJWT==2.6.0
pypng==0.20220715.0
pytz==2022.7.1
qrcode==7.4.2
sqlparse==0.4.3
typing_extensions==4.7.1
